requirements.txt
flask
google-cloud-bigquery
functions-framework
